move(cosmo): move connect to private API
