+++ astropy/cosmology/_src/io/__init__.py
@@ -1,0 +1,5 @@
+"""Astropy Cosmology IO. **NOT public API**.
+
+The public API is provided by `astropy.cosmology.io`.
+"""
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
